chatbot_backend/requirements.txt
Flask 
Flask-CORS 
python-dotenv
 Werkzeug
SQLAlchemy
Flask-SQLAlchemy
pymysql
cryptography